add unit-tests for  noise protocols
